chore(tests): move ParseNestedTableExtraFields to plugin_collection
