+++ falco_plugin_tests/src/plugin_collection/parse/mod.rs
@@ -1,4 +1,5 @@
 pub mod extra_fields;
+pub mod nested_table_extra_fields;
 pub mod remaining_into_nested_table;
 pub mod remaining_into_table_api;
 pub mod remaining_into_table_direct;
